easyblue_backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
flask-cors
psycopg2-binary
python-dotenv
sentry-sdk[flask]
